Move macOS DevMenuWindow implementation to a separete file
