+++ packages/expo-dev-launcher/ios/SwiftUI/Navigation/Navigation.swift
@@ -1,6 +1,5 @@
 // swiftlint:disable closure_body_length
 import SwiftUI
-import ExpoModulesCore
 
 class DevLauncherNavigation: ObservableObject {
   @Binding var showingUserProfile: Bool
